+++ include/arch/gb/console.h
@@ -10,7 +10,7 @@
 
 /** Move the cursor to an absolute position.
  */
-void __LIB__ gotoxy(uint8_t x, uint8_t y) __smallc;
+void __LIB__ gotoxy(uint16_t x, uint16_t y) __smallc;
 
 /** Get the current X position of the cursor.
  */
